Final-year-project-master/server/requirements.txt
absl-py
anyio
astunparse
cachetools
certifi
charset-normalizer
click
colorama
cycler
fastapi
Flask
Flask-Cors
flatbuffers
fonttools
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h11
h5py
idna
importlib-metadata
itsdangerous
Jinja2
keras
kiwisolver
libclang
Markdown
MarkupSafe
matplotlib
numpy
oauthlib
opencv-python
opt-einsum
packaging
pandas
Pillow
protobuf==3.20.1
pyasn1
pyasn1-modules
pydantic
pyparsing
python-dateutil
python-multipart
pytz
requests
requests-oauthlib
rsa
scipy
six
sniffio
starlette
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
typing_extensions
urllib3
uvicorn
Werkzeug
wrapt
xgboost
zipp